Feature: Implementação do CRUD de Address e Ongs

Aprovado

+++ src/main/java/com/adotai/backend_adotai/dto/User/request/RequestUserDTO.java
@@ -11,5 +11,5 @@ public record RequestUserDTO(String name,
                              String password,
                              String telephone,
                              String role,
-                             Long addressId){}
+                             int addressId){}
